Exception handling

+++ src/main/kotlin/org/example/dscommerce/services/ProductService.kt
@@ -1,12 +1,17 @@
 package org.example.dscommerce.services
 
+
 import org.example.dscommerce.dto.ProductDTO
 import org.example.dscommerce.entities.Product
 import org.example.dscommerce.mappers.toDTO
 import org.example.dscommerce.mappers.toEntity
 import org.example.dscommerce.repositories.ProductRepository
+import org.example.dscommerce.services.exceptions.ResourceNotFoundException
+import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
+import org.springframework.data.repository.findByIdOrNull
+import org.springframework.orm.ObjectRetrievalFailureException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -15,8 +20,9 @@ class ProductService(private val productRepository: ProductRepository) {
 
     @Transactional(readOnly = true)
     fun findById(id: Long): ProductDTO {
-        val result = productRepository.findById(id)
-        val product = result.get()
+        val product = productRepository.findById(id).orElseThrow {
+            ResourceNotFoundException("Product with ID $id not found")
+        }
         return product.toDTO()
     }
 
@@ -28,20 +34,27 @@ class ProductService(private val productRepository: ProductRepository) {
     }
 
 
-    @Transactional fun insert(productDTO: ProductDTO): ProductDTO {
+    @Transactional
+    fun insert(productDTO: ProductDTO): ProductDTO {
         val newProduct = productDTO.toEntity()
         val result = productRepository.save(newProduct)
         return result.toDTO()
     }
 
-    @Transactional fun update(id: Long, productDTO: ProductDTO): ProductDTO {
-        val product = productRepository.getReferenceById(id)
-        val updatedProduct = Product(product.id, productDTO.name, productDTO.description, productDTO.price, productDTO.imgUrl)
+    @Transactional
+    fun update(id: Long, productDTO: ProductDTO): ProductDTO {
+        val product = productRepository.findByIdOrNull(id) ?: throw ResourceNotFoundException("Product with ID $id not found")
+        val updatedProduct =
+            Product(product.id, productDTO.name, productDTO.description, productDTO.price, productDTO.imgUrl)
         val savedProduct = productRepository.save(updatedProduct)
         return savedProduct.toDTO()
     }
 
-    @Transactional fun delete(id: Long) {
+    @Transactional
+    fun delete(id: Long) {
+        if (!productRepository.existsById(id)) {
+            throw ResourceNotFoundException("Product with ID $id not found")
+        }
         productRepository.deleteById(id)
     }
